Move CalculableState to Extension
