+++ src/Extension/CustomStateFoundation.php
@@ -4,7 +4,7 @@
 namespace Wirecard\Order\State\Extension;
 
 
-use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Extension\CalculableState;
 use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
 use Wirecard\Order\State\State;
