+++ src/State/Accepted.php
@@ -3,15 +3,13 @@
 
 namespace Wirecard\Order\State\State;
 
-use Wirecard\Order\State\Implementation\CalculableState;
-use Wirecard\Order\State\State\Failed;
+use Wirecard\Order\State\Extension\CalculableState;
 use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
-use Wirecard\Order\State\State;
 
 class Accepted implements CalculableState
 {
-    use \Wirecard\Order\State\Implementation\StateHelper;
+    use StateHelper;
 
     public function getPossibleNextStates()
     {
