+++ src/State/Authorized.php
@@ -3,10 +3,9 @@
 
 namespace Wirecard\Order\State\State;
 
-use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Extension\CalculableState;
 use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
-use Wirecard\Order\State\State;
 
 class Authorized implements CalculableState
 {
