+++ src/State/Processing.php
@@ -3,10 +3,9 @@
 
 namespace Wirecard\Order\State\State;
 
-use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Extension\CalculableState;
 use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
-use Wirecard\Order\State\State\Failed;
 
 /**
  * Class Processing
@@ -18,7 +17,7 @@ use Wirecard\Order\State\State\Failed;
  */
 class Processing implements CalculableState
 {
-    use \Wirecard\Order\State\Implementation\StateHelper;
+    use StateHelper;
 
     public function getPossibleNextStates()
     {
