+++ src/State/Success.php
@@ -5,15 +5,13 @@ namespace Wirecard\Order\State\State;
 
 
 use Wirecard\Order\State\CreditCardTransactionType\PurchaseTransaction;
-use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Extension\CalculableState;
 use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
-use Wirecard\Order\State\State\Authorized;
-use Wirecard\Order\State\State\Processing;
 
-class Success implements \Wirecard\Order\State\Implementation\CalculableState
+class Success implements CalculableState
 {
-    use \Wirecard\Order\State\Implementation\StateHelper;
+    use StateHelper;
 
     public function getPossibleNextStates()
     {
